+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version: 4.1.1
 ENHANCEMENTS:
 
 * Improvement done when reporting exceptions, as despite the traceback is being reported, the exception message was not being displayed.
+* Improvement done to display the correct exit code when execution crashes in environment hooks.
 
 
 Version: 4.1.0